added security when u delete any profile it request to enter padssword done which is running localhost:3000

+++ src/components/students/Students.js
@@ -3,6 +3,8 @@ import { Link } from 'react-router-dom'
 import Avatar from '../layout/Avatar'
 import { useFirestoreConnect,useFirestore } from "react-redux-firebase";
 import {useSelector} from 'react-redux';
+import img from '../../assets/22.jpg';
+
 
 
 const Students = () => {
@@ -18,12 +20,17 @@ const Students = () => {
     return <h1>Loading...</h1>
   }
 
-    const deleteStudent = async (id) =>{
-      try {
-        await firestore.collection("students").doc(id).delete();
-      } catch (error) {
-        alert("Something Wrong!!");                                                               
-      }
+    const deleteStudent = async (id,standard) =>{
+        let pass = prompt("Enter password");
+        if(pass===(standard))
+        {
+          await firestore.collection("students").doc(id).delete();
+
+        }
+        else{
+          alert("Something Wrong");                                                               
+        }
+
     }       
 
     return (
@@ -34,14 +41,14 @@ const Students = () => {
           students.map((student)=>(
             <div className="col-lg-3 col-md-6 mb-4" key={student.id}>
   <div className="card shadow text-center py-4">
-  <Avatar url={`https://i.pravatar.cc/150?img=${student.id}`} />
+  <Avatar url={img} />
     <div className="card-body">
       <h5 className="card-title mb-0">{student.name}</h5>
       <p className="text-muted small">{student.email}</p>
       <Link to={`/student/${student.id}`} className="btn btn-primary btn-profile">
         View Profile
       </Link>
-      <button className="btn btn-edit" onClick={()=>deleteStudent(student.id)}>
+      <button className="btn btn-edit" onClick={()=>deleteStudent(student.id,student.standard)}>
         <span className="material-icons">delete_outline</span>
       </button>
     </div>
